Add thumbnailImage Helper
Given a well-formed object with values for an image, will return a
`figure` with an `img` (optionally wrapped in a link), and an optional
`figcaption`.

Likely usecase is hooking the given class `js-thumbnail` (on figure)
to lightbox script of your choice.

For a gallery, the recommendation is to wrap thumbnails in a list.