add concat helper for array type
concat uses Array.prototype.concat to concatinate two arrays and return 
the result.

In spirit of the design of other helpers, I made it so that if only the 
first argument is an array, that array will be returned. If both 
arguments are an array their concatination is returned. If neither 
arguments are an array then an empty array is returned.