Add fast-fails to Array library helpers
Array helpers.forEach was causing errors that resulted in status 500 errors in runtime when passed undefined or null values. Added `if (!Array.isArray(array)) return '';`  as a fast-fail measure to it and other Array helpers to handle bad parameters without causing errors. This fast-fail was already being used on multiple other helpers in this file, and was copied and added to those which might error without it.